move workspace tests
