+++ test/cli/install/bun-workspaces.test.ts
@@ -1,29 +1,49 @@
-import { file, write } from "bun";
+import { file, write, spawn } from "bun";
 import { install_test_helpers } from "bun:internal-for-testing";
-import { beforeEach, describe, expect, test } from "bun:test";
+import { beforeEach, describe, expect, test, beforeAll, afterAll } from "bun:test";
 import { mkdirSync, rmSync, writeFileSync } from "fs";
-import { cp } from "fs/promises";
-import { bunExe, bunEnv as env, runBunInstall, tmpdirSync, toMatchNodeModulesAt } from "harness";
+import { cp, mkdir, rm, exists } from "fs/promises";
+import {
+  bunExe,
+  bunEnv as env,
+  runBunInstall,
+  tmpdirSync,
+  toMatchNodeModulesAt,
+  assertManifestsPopulated,
+  VerdaccioRegistry,
+  readdirSorted,
+} from "harness";
 import { join } from "path";
 const { parseLockfile } = install_test_helpers;
 
 expect.extend({ toMatchNodeModulesAt });
 
-var testCounter: number = 0;
-
 // not necessary, but verdaccio will be added to this file in the near future
-var port: number = 4873;
+
+var verdaccio: VerdaccioRegistry;
 var packageDir: string;
+var packageJson: string;
+
+beforeAll(async () => {
+  verdaccio = new VerdaccioRegistry();
+  await verdaccio.start();
+});
+
+afterAll(() => {
+  verdaccio.stop();
+});
 
 beforeEach(() => {
   packageDir = tmpdirSync();
+  packageJson = join(packageDir, "package.json");
   env.BUN_INSTALL_CACHE_DIR = join(packageDir, ".bun-cache");
   env.BUN_TMPDIR = env.TMPDIR = env.TEMP = join(packageDir, ".bun-tmp");
   writeFileSync(
     join(packageDir, "bunfig.toml"),
     `
 [install]
-cache = false
+cache = "${join(packageDir, ".bun-cache")}"
+registry = "http://localhost:${verdaccio.port}/"
 `,
   );
 });
@@ -41,7 +61,7 @@ test("dependency on workspace without version in package.json", async () => {
     write(
       join(packageDir, "packages", "mono", "package.json"),
       JSON.stringify({
-        name: "lodash",
+        name: "no-deps",
       }),
     ),
   ]);
@@ -60,7 +80,7 @@ test("dependency on workspace without version in package.json", async () => {
     "1",
     "1.*",
     "1.1.*",
-    "1.1.1",
+    "1.1.0",
     "*-pre+build",
     "*+build",
     "latest", // dist-tag exists, should choose package from npm
@@ -74,7 +94,7 @@ test("dependency on workspace without version in package.json", async () => {
         name: "bar",
         version: "1.0.0",
         dependencies: {
-          lodash: version,
+          "no-deps": version,
         },
       }),
     );
@@ -82,7 +102,9 @@ test("dependency on workspace without version in package.json", async () => {
     const { out } = await runBunInstall(env, packageDir);
     const lockfile = parseLockfile(packageDir);
     expect(lockfile).toMatchNodeModulesAt(packageDir);
-    expect(lockfile).toMatchSnapshot(`version: ${version}`);
+    expect(
+      JSON.stringify(lockfile, null, 2).replaceAll(/http:\/\/localhost:\d+/g, "http://localhost:1234"),
+    ).toMatchSnapshot(`version: ${version}`);
     expect(out.replace(/\s*\[[0-9\.]+m?s\]\s*$/, "").split(/\r?\n/)).toEqual([
       expect.stringContaining("bun install v1."),
       "",
@@ -101,7 +123,7 @@ test("dependency on workspace without version in package.json", async () => {
         name: "bar",
         version: "1.0.0",
         dependencies: {
-          lodash: version,
+          "no-deps": version,
         },
       }),
     );
@@ -109,7 +131,9 @@ test("dependency on workspace without version in package.json", async () => {
     const { out } = await runBunInstall(env, packageDir);
     const lockfile = parseLockfile(packageDir);
     expect(lockfile).toMatchNodeModulesAt(packageDir);
-    expect(lockfile).toMatchSnapshot(`version: ${version}`);
+    expect(
+      JSON.stringify(lockfile, null, 2).replaceAll(/http:\/\/localhost:\d+/g, "http://localhost:1234"),
+    ).toMatchSnapshot(`version: ${version}`);
     expect(out.replace(/\s*\[[0-9\.]+m?s\]\s*$/, "").split(/\r?\n/)).toEqual([
       expect.stringContaining("bun install v1."),
       "",
@@ -134,7 +158,7 @@ test("dependency on same name as workspace and dist-tag", async () => {
     write(
       join(packageDir, "packages", "mono", "package.json"),
       JSON.stringify({
-        name: "lodash",
+        name: "no-deps",
         version: "4.17.21",
       }),
     ),
@@ -145,7 +169,7 @@ test("dependency on same name as workspace and dist-tag", async () => {
         name: "bar",
         version: "1.0.0",
         dependencies: {
-          lodash: "latest",
+          "no-deps": "latest",
         },
       }),
     ),
@@ -153,7 +177,9 @@ test("dependency on same name as workspace and dist-tag", async () => {
 
   const { out } = await runBunInstall(env, packageDir);
   const lockfile = parseLockfile(packageDir);
-  expect(lockfile).toMatchSnapshot("with version");
+  expect(
+    JSON.stringify(lockfile, null, 2).replaceAll(/http:\/\/localhost:\d+/g, "http://localhost:1234"),
+  ).toMatchSnapshot("with version");
   expect(lockfile).toMatchNodeModulesAt(packageDir);
   expect(out.replace(/\s*\[[0-9\.]+m?s\]\s*$/, "").split(/\r?\n/)).toEqual([
     expect.stringContaining("bun install v1."),
@@ -658,3 +684,611 @@ test("$npm_package_config_ works in root in subpackage", async () => {
   expect(await new Response(p.stderr).text()).toBe(`$ echo $npm_package_config_foo $npm_package_config_qux\n`);
   expect(await new Response(p.stdout).text()).toBe(`tab\n`);
 });
+
+test("adding packages in a subdirectory of a workspace", async () => {
+  await write(
+    packageJson,
+    JSON.stringify({
+      name: "root",
+      workspaces: ["foo"],
+    }),
+  );
+
+  await mkdir(join(packageDir, "folder1"));
+  await mkdir(join(packageDir, "foo", "folder2"), { recursive: true });
+  await write(
+    join(packageDir, "foo", "package.json"),
+    JSON.stringify({
+      name: "foo",
+    }),
+  );
+
+  // add package to root workspace from `folder1`
+  let { stdout, exited } = spawn({
+    cmd: [bunExe(), "add", "no-deps"],
+    cwd: join(packageDir, "folder1"),
+    stdout: "pipe",
+    stderr: "inherit",
+    env,
+  });
+  let out = await Bun.readableStreamToText(stdout);
+  expect(out.replace(/\s*\[[0-9\.]+m?s\]\s*$/, "").split(/\r?\n/)).toEqual([
+    expect.stringContaining("bun add v1."),
+    "",
+    "installed no-deps@2.0.0",
+    "",
+    "2 packages installed",
+  ]);
+  expect(await exited).toBe(0);
+  assertManifestsPopulated(join(packageDir, ".bun-cache"), verdaccio.registryUrl());
+
+  expect(await file(packageJson).json()).toEqual({
+    name: "root",
+    workspaces: ["foo"],
+    dependencies: {
+      "no-deps": "^2.0.0",
+    },
+  });
+
+  // add package to foo from `folder2`
+  ({ stdout, exited } = spawn({
+    cmd: [bunExe(), "add", "what-bin"],
+    cwd: join(packageDir, "foo", "folder2"),
+    stdout: "pipe",
+    stderr: "inherit",
+    env,
+  }));
+  out = await Bun.readableStreamToText(stdout);
+  expect(out.replace(/\s*\[[0-9\.]+m?s\]\s*$/, "").split(/\r?\n/)).toEqual([
+    expect.stringContaining("bun add v1."),
+    "",
+    "installed what-bin@1.5.0 with binaries:",
+    " - what-bin",
+    "",
+    "1 package installed",
+  ]);
+  expect(await exited).toBe(0);
+  assertManifestsPopulated(join(packageDir, ".bun-cache"), verdaccio.registryUrl());
+
+  expect(await file(join(packageDir, "foo", "package.json")).json()).toEqual({
+    name: "foo",
+    dependencies: {
+      "what-bin": "^1.5.0",
+    },
+  });
+
+  // now delete node_modules and bun.lockb and install
+  await rm(join(packageDir, "node_modules"), { recursive: true, force: true });
+  await rm(join(packageDir, "bun.lockb"));
+
+  ({ stdout, exited } = spawn({
+    cmd: [bunExe(), "install"],
+    cwd: join(packageDir, "folder1"),
+    stdout: "pipe",
+    stderr: "inherit",
+    env,
+  }));
+  out = await Bun.readableStreamToText(stdout);
+  expect(out.replace(/\s*\[[0-9\.]+m?s\]\s*$/, "").split(/\r?\n/)).toEqual([
+    expect.stringContaining("bun install v1."),
+    "",
+    "+ no-deps@2.0.0",
+    "",
+    "3 packages installed",
+  ]);
+  expect(await exited).toBe(0);
+  assertManifestsPopulated(join(packageDir, ".bun-cache"), verdaccio.registryUrl());
+
+  expect(await readdirSorted(join(packageDir, "node_modules"))).toEqual([".bin", "foo", "no-deps", "what-bin"]);
+
+  await rm(join(packageDir, "node_modules"), { recursive: true, force: true });
+  await rm(join(packageDir, "bun.lockb"));
+
+  ({ stdout, exited } = spawn({
+    cmd: [bunExe(), "install"],
+    cwd: join(packageDir, "foo", "folder2"),
+    stdout: "pipe",
+    stderr: "inherit",
+    env,
+  }));
+  out = await Bun.readableStreamToText(stdout);
+  expect(out.replace(/\s*\[[0-9\.]+m?s\]\s*$/, "").split(/\r?\n/)).toEqual([
+    expect.stringContaining("bun install v1."),
+    "",
+    "+ what-bin@1.5.0",
+    "",
+    "3 packages installed",
+  ]);
+  expect(await exited).toBe(0);
+  assertManifestsPopulated(join(packageDir, ".bun-cache"), verdaccio.registryUrl());
+
+  expect(await readdirSorted(join(packageDir, "node_modules"))).toEqual([".bin", "foo", "no-deps", "what-bin"]);
+});
+test("adding packages in workspaces", async () => {
+  await write(
+    packageJson,
+    JSON.stringify({
+      name: "foo",
+      workspaces: ["packages/*"],
+      dependencies: {
+        "bar": "workspace:*",
+      },
+    }),
+  );
+
+  await mkdir(join(packageDir, "packages", "bar"), { recursive: true });
+  await mkdir(join(packageDir, "packages", "boba"));
+  await mkdir(join(packageDir, "packages", "pkg5"));
+
+  await write(join(packageDir, "packages", "bar", "package.json"), JSON.stringify({ name: "bar" }));
+  await write(
+    join(packageDir, "packages", "boba", "package.json"),
+    JSON.stringify({ name: "boba", version: "1.0.0", dependencies: { "pkg5": "*" } }),
+  );
+  await write(
+    join(packageDir, "packages", "pkg5", "package.json"),
+    JSON.stringify({
+      name: "pkg5",
+      version: "1.2.3",
+      dependencies: {
+        "bar": "workspace:*",
+      },
+    }),
+  );
+
+  let { stdout, exited } = spawn({
+    cmd: [bunExe(), "install"],
+    cwd: packageDir,
+    stdout: "pipe",
+    stderr: "inherit",
+    env,
+  });
+
+  let out = await Bun.readableStreamToText(stdout);
+  expect(out.replace(/\s*\[[0-9\.]+m?s\]\s*$/, "").split(/\r?\n/)).toEqual([
+    expect.stringContaining("bun install v1."),
+    "",
+    "+ bar@workspace:packages/bar",
+    "",
+    "3 packages installed",
+  ]);
+  expect(await exited).toBe(0);
+  assertManifestsPopulated(join(packageDir, ".bun-cache"), verdaccio.registryUrl());
+
+  expect(await exists(join(packageDir, "node_modules", "bar"))).toBeTrue();
+  expect(await exists(join(packageDir, "node_modules", "boba"))).toBeTrue();
+  expect(await exists(join(packageDir, "node_modules", "pkg5"))).toBeTrue();
+
+  // add a package to the root workspace
+  ({ stdout, exited } = spawn({
+    cmd: [bunExe(), "add", "no-deps"],
+    cwd: packageDir,
+    stdout: "pipe",
+    stderr: "inherit",
+    env,
+  }));
+
+  out = await Bun.readableStreamToText(stdout);
+  expect(out.replace(/\s*\[[0-9\.]+m?s\]\s*$/, "").split(/\r?\n/)).toEqual([
+    expect.stringContaining("bun add v1."),
+    "",
+    "installed no-deps@2.0.0",
+    "",
+    "1 package installed",
+  ]);
+  expect(await exited).toBe(0);
+  assertManifestsPopulated(join(packageDir, ".bun-cache"), verdaccio.registryUrl());
+
+  expect(await file(packageJson).json()).toEqual({
+    name: "foo",
+    workspaces: ["packages/*"],
+    dependencies: {
+      bar: "workspace:*",
+      "no-deps": "^2.0.0",
+    },
+  });
+
+  // add a package in a workspace
+  ({ stdout, exited } = spawn({
+    cmd: [bunExe(), "add", "two-range-deps"],
+    cwd: join(packageDir, "packages", "boba"),
+    stdout: "pipe",
+    stderr: "inherit",
+    env,
+  }));
+
+  out = await Bun.readableStreamToText(stdout);
+  expect(out.replace(/\s*\[[0-9\.]+m?s\]\s*$/, "").split(/\r?\n/)).toEqual([
+    expect.stringContaining("bun add v1."),
+    "",
+    "installed two-range-deps@1.0.0",
+    "",
+    "3 packages installed",
+  ]);
+  expect(await exited).toBe(0);
+  assertManifestsPopulated(join(packageDir, ".bun-cache"), verdaccio.registryUrl());
+
+  expect(await file(join(packageDir, "packages", "boba", "package.json")).json()).toEqual({
+    name: "boba",
+    version: "1.0.0",
+    dependencies: {
+      "pkg5": "*",
+      "two-range-deps": "^1.0.0",
+    },
+  });
+  expect(await readdirSorted(join(packageDir, "node_modules"))).toEqual([
+    "@types",
+    "bar",
+    "boba",
+    "no-deps",
+    "pkg5",
+    "two-range-deps",
+  ]);
+
+  // add a dependency to a workspace with the same name as another workspace
+  ({ stdout, exited } = spawn({
+    cmd: [bunExe(), "add", "bar@0.0.7"],
+    cwd: join(packageDir, "packages", "boba"),
+    stdout: "pipe",
+    stderr: "inherit",
+    env,
+  }));
+
+  out = await Bun.readableStreamToText(stdout);
+  expect(out.replace(/\s*\[[0-9\.]+m?s\]\s*$/, "").split(/\r?\n/)).toEqual([
+    expect.stringContaining("bun add v1."),
+    "",
+    "installed bar@0.0.7",
+    "",
+    "1 package installed",
+  ]);
+  expect(await exited).toBe(0);
+  assertManifestsPopulated(join(packageDir, ".bun-cache"), verdaccio.registryUrl());
+
+  expect(await file(join(packageDir, "packages", "boba", "package.json")).json()).toEqual({
+    name: "boba",
+    version: "1.0.0",
+    dependencies: {
+      "pkg5": "*",
+      "two-range-deps": "^1.0.0",
+      "bar": "0.0.7",
+    },
+  });
+  expect(await readdirSorted(join(packageDir, "node_modules"))).toEqual([
+    "@types",
+    "bar",
+    "boba",
+    "no-deps",
+    "pkg5",
+    "two-range-deps",
+  ]);
+  expect(await file(join(packageDir, "node_modules", "boba", "node_modules", "bar", "package.json")).json()).toEqual({
+    name: "bar",
+    version: "0.0.7",
+    description: "not a workspace",
+  });
+});
+test("it should detect duplicate workspace dependencies", async () => {
+  await write(
+    packageJson,
+    JSON.stringify({
+      name: "foo",
+      workspaces: ["packages/*"],
+    }),
+  );
+
+  await mkdir(join(packageDir, "packages", "pkg1"), { recursive: true });
+  await write(join(packageDir, "packages", "pkg1", "package.json"), JSON.stringify({ name: "pkg1" }));
+  await mkdir(join(packageDir, "packages", "pkg2"), { recursive: true });
+  await write(join(packageDir, "packages", "pkg2", "package.json"), JSON.stringify({ name: "pkg1" }));
+
+  var { stderr, exited } = spawn({
+    cmd: [bunExe(), "install"],
+    cwd: packageDir,
+    stdout: "pipe",
+    stdin: "pipe",
+    stderr: "pipe",
+    env,
+  });
+
+  var err = await new Response(stderr).text();
+  expect(err).toContain('Workspace name "pkg1" already exists');
+  expect(await exited).toBe(1);
+
+  await rm(join(packageDir, "node_modules"), { recursive: true, force: true });
+  await rm(join(packageDir, "bun.lockb"), { force: true });
+
+  ({ stderr, exited } = spawn({
+    cmd: [bunExe(), "install"],
+    cwd: join(packageDir, "packages", "pkg1"),
+    stdout: "pipe",
+    stdin: "pipe",
+    stderr: "pipe",
+    env,
+  }));
+
+  err = await new Response(stderr).text();
+  expect(err).toContain('Workspace name "pkg1" already exists');
+  expect(await exited).toBe(1);
+});
+
+const versions = ["workspace:1.0.0", "workspace:*", "workspace:^1.0.0", "1.0.0", "*"];
+
+for (const rootVersion of versions) {
+  for (const packageVersion of versions) {
+    test(`it should allow duplicates, root@${rootVersion}, package@${packageVersion}`, async () => {
+      await write(
+        packageJson,
+        JSON.stringify({
+          name: "foo",
+          version: "1.0.0",
+          workspaces: ["packages/*"],
+          dependencies: {
+            pkg2: rootVersion,
+          },
+        }),
+      );
+
+      await mkdir(join(packageDir, "packages", "pkg1"), { recursive: true });
+      await write(
+        join(packageDir, "packages", "pkg1", "package.json"),
+        JSON.stringify({
+          name: "pkg1",
+          version: "1.0.0",
+          dependencies: {
+            pkg2: packageVersion,
+          },
+        }),
+      );
+
+      await mkdir(join(packageDir, "packages", "pkg2"), { recursive: true });
+      await write(
+        join(packageDir, "packages", "pkg2", "package.json"),
+        JSON.stringify({ name: "pkg2", version: "1.0.0" }),
+      );
+
+      var { stdout, stderr, exited } = spawn({
+        cmd: [bunExe(), "install"],
+        cwd: packageDir,
+        stdout: "pipe",
+        stdin: "pipe",
+        stderr: "pipe",
+        env,
+      });
+
+      var err = await new Response(stderr).text();
+      var out = await new Response(stdout).text();
+      expect(err).toContain("Saved lockfile");
+      expect(err).not.toContain("not found");
+      expect(err).not.toContain("error:");
+      expect(out.replace(/\s*\[[0-9\.]+m?s\]\s*$/, "").split(/\r?\n/)).toEqual([
+        expect.stringContaining("bun install v1."),
+        "",
+        `+ pkg2@workspace:packages/pkg2`,
+        "",
+        "2 packages installed",
+      ]);
+      expect(await exited).toBe(0);
+      assertManifestsPopulated(join(packageDir, ".bun-cache"), verdaccio.registryUrl());
+
+      ({ stdout, stderr, exited } = spawn({
+        cmd: [bunExe(), "install"],
+        cwd: join(packageDir, "packages", "pkg1"),
+        stdout: "pipe",
+        stdin: "pipe",
+        stderr: "pipe",
+        env,
+      }));
+
+      err = await new Response(stderr).text();
+      out = await new Response(stdout).text();
+      expect(err).not.toContain("Saved lockfile");
+      expect(err).not.toContain("not found");
+      expect(err).not.toContain("error:");
+      expect(out.replace(/\s*\[[0-9\.]+m?s\]\s*$/, "").split(/\r?\n/)).toEqual([
+        expect.stringContaining("bun install v1."),
+        "",
+        "Checked 2 installs across 3 packages (no changes)",
+      ]);
+      expect(await exited).toBe(0);
+      assertManifestsPopulated(join(packageDir, ".bun-cache"), verdaccio.registryUrl());
+
+      await rm(join(packageDir, "node_modules"), { recursive: true, force: true });
+      await rm(join(packageDir, "bun.lockb"), { recursive: true, force: true });
+
+      ({ stdout, stderr, exited } = spawn({
+        cmd: [bunExe(), "install"],
+        cwd: join(packageDir, "packages", "pkg1"),
+        stdout: "pipe",
+        stdin: "pipe",
+        stderr: "pipe",
+        env,
+      }));
+
+      err = await new Response(stderr).text();
+      out = await new Response(stdout).text();
+      expect(err).toContain("Saved lockfile");
+      expect(err).not.toContain("not found");
+      expect(err).not.toContain("error:");
+      expect(out.replace(/\s*\[[0-9\.]+m?s\]\s*$/, "").split(/\r?\n/)).toEqual([
+        expect.stringContaining("bun install v1."),
+        "",
+        `+ pkg2@workspace:packages/pkg2`,
+        "",
+        "2 packages installed",
+      ]);
+      expect(await exited).toBe(0);
+      assertManifestsPopulated(join(packageDir, ".bun-cache"), verdaccio.registryUrl());
+
+      ({ stdout, stderr, exited } = spawn({
+        cmd: [bunExe(), "install"],
+        cwd: packageDir,
+        stdout: "pipe",
+        stdin: "pipe",
+        stderr: "pipe",
+        env,
+      }));
+
+      err = await new Response(stderr).text();
+      out = await new Response(stdout).text();
+      expect(err).not.toContain("Saved lockfile");
+      expect(err).not.toContain("not found");
+      expect(err).not.toContain("error:");
+      expect(out.replace(/\s*\[[0-9\.]+m?s\]\s*$/, "").split(/\r?\n/)).toEqual([
+        expect.stringContaining("bun install v1."),
+        "",
+        "Checked 2 installs across 3 packages (no changes)",
+      ]);
+      expect(await exited).toBe(0);
+      assertManifestsPopulated(join(packageDir, ".bun-cache"), verdaccio.registryUrl());
+    });
+  }
+}
+
+for (const version of versions) {
+  test(`it should allow listing workspace as dependency of the root package version ${version}`, async () => {
+    await write(
+      packageJson,
+      JSON.stringify({
+        name: "foo",
+        workspaces: ["packages/*"],
+        dependencies: {
+          "workspace-1": version,
+        },
+      }),
+    );
+
+    await mkdir(join(packageDir, "packages", "workspace-1"), { recursive: true });
+    await write(
+      join(packageDir, "packages", "workspace-1", "package.json"),
+      JSON.stringify({
+        name: "workspace-1",
+        version: "1.0.0",
+      }),
+    );
+    // install first from the root, the workspace package
+    var { stdout, stderr, exited } = spawn({
+      cmd: [bunExe(), "install"],
+      cwd: packageDir,
+      stdout: "pipe",
+      stdin: "pipe",
+      stderr: "pipe",
+      env,
+    });
+
+    var err = await new Response(stderr).text();
+    var out = await new Response(stdout).text();
+    expect(err).toContain("Saved lockfile");
+    expect(err).not.toContain("already exists");
+    expect(err).not.toContain("not found");
+    expect(err).not.toContain("Duplicate dependency");
+    expect(err).not.toContain('workspace dependency "workspace-1" not found');
+    expect(err).not.toContain("error:");
+    expect(out.replace(/\s*\[[0-9\.]+m?s\]\s*$/, "").split(/\r?\n/)).toEqual([
+      expect.stringContaining("bun install v1."),
+      "",
+      `+ workspace-1@workspace:packages/workspace-1`,
+      "",
+      "1 package installed",
+    ]);
+    expect(await exited).toBe(0);
+    assertManifestsPopulated(join(packageDir, ".bun-cache"), verdaccio.registryUrl());
+
+    expect(await file(join(packageDir, "node_modules", "workspace-1", "package.json")).json()).toEqual({
+      name: "workspace-1",
+      version: "1.0.0",
+    });
+
+    ({ stdout, stderr, exited } = spawn({
+      cmd: [bunExe(), "install"],
+      cwd: join(packageDir, "packages", "workspace-1"),
+      stdout: "pipe",
+      stdin: "pipe",
+      stderr: "pipe",
+      env,
+    }));
+
+    err = await new Response(stderr).text();
+    out = await new Response(stdout).text();
+    expect(err).not.toContain("Saved lockfile");
+    expect(err).not.toContain("not found");
+    expect(err).not.toContain("already exists");
+    expect(err).not.toContain("Duplicate dependency");
+    expect(err).not.toContain('workspace dependency "workspace-1" not found');
+    expect(err).not.toContain("error:");
+    expect(out.replace(/\s*\[[0-9\.]+m?s\]\s*$/, "").split(/\r?\n/)).toEqual([
+      expect.stringContaining("bun install v1."),
+      "",
+      "Checked 1 install across 2 packages (no changes)",
+    ]);
+    expect(await exited).toBe(0);
+    assertManifestsPopulated(join(packageDir, ".bun-cache"), verdaccio.registryUrl());
+
+    expect(await file(join(packageDir, "node_modules", "workspace-1", "package.json")).json()).toEqual({
+      name: "workspace-1",
+      version: "1.0.0",
+    });
+
+    await rm(join(packageDir, "node_modules"), { recursive: true, force: true });
+    await rm(join(packageDir, "bun.lockb"), { recursive: true, force: true });
+
+    // install from workspace package then from root
+    ({ stdout, stderr, exited } = spawn({
+      cmd: [bunExe(), "install"],
+      cwd: join(packageDir, "packages", "workspace-1"),
+      stdout: "pipe",
+      stdin: "pipe",
+      stderr: "pipe",
+      env,
+    }));
+
+    err = await new Response(stderr).text();
+    out = await new Response(stdout).text();
+    expect(err).toContain("Saved lockfile");
+    expect(err).not.toContain("already exists");
+    expect(err).not.toContain("not found");
+    expect(err).not.toContain("Duplicate dependency");
+    expect(err).not.toContain('workspace dependency "workspace-1" not found');
+    expect(err).not.toContain("error:");
+    expect(out.replace(/\s*\[[0-9\.]+m?s\]\s*$/, "").split(/\r?\n/)).toEqual([
+      expect.stringContaining("bun install v1."),
+      "",
+      "1 package installed",
+    ]);
+    expect(await exited).toBe(0);
+    expect(await file(join(packageDir, "node_modules", "workspace-1", "package.json")).json()).toEqual({
+      name: "workspace-1",
+      version: "1.0.0",
+    });
+
+    ({ stdout, stderr, exited } = spawn({
+      cmd: [bunExe(), "install"],
+      cwd: packageDir,
+      stdout: "pipe",
+      stdin: "pipe",
+      stderr: "pipe",
+      env,
+    }));
+
+    err = await new Response(stderr).text();
+    out = await new Response(stdout).text();
+    expect(err).not.toContain("Saved lockfile");
+    expect(err).not.toContain("already exists");
+    expect(err).not.toContain("not found");
+    expect(err).not.toContain("Duplicate dependency");
+    expect(err).not.toContain('workspace dependency "workspace-1" not found');
+    expect(err).not.toContain("error:");
+    expect(out.replace(/\s*\[[0-9\.]+m?s\]\s*$/, "").split(/\r?\n/)).toEqual([
+      expect.stringContaining("bun install v1."),
+      "",
+      "Checked 1 install across 2 packages (no changes)",
+    ]);
+    expect(await exited).toBe(0);
+    assertManifestsPopulated(join(packageDir, ".bun-cache"), verdaccio.registryUrl());
+
+    expect(await file(join(packageDir, "node_modules", "workspace-1", "package.json")).json()).toEqual({
+      name: "workspace-1",
+      version: "1.0.0",
+    });
+  });
+}
